move is_openai_reasoning_model to openai.rs
